[Refactor] #439 스케줄러 락 설정을 common으로 올리고 @Import로 활성화한다
booking·seat·performance에 동일 FQN으로 복붙돼 있던 ShedLockConfig를 common
한 곳으로 옮긴다. 세 파일의 실질 차이는 @Value 기본값 문자열 한 줄이었고, 그
기본값조차 세 모듈 모두 application.yml에 이름을 정의해 둬 도달하지 않았다.

common의 @Configuration은 전 서비스에 컴포넌트 스캔되므로, 그대로 옮기면 ShedLock을
쓰지 않는 payment·user·ticket·auth까지 이 설정을 로드한다. 그래서 ShedLockConfig에는
@Configuration을 붙이지 않아 스캔 대상에서 빼고, 락이 필요한 서비스가 각자
SchedulingConfig에서 @Import로 명시 활성화한다.

기존 선례인 RedisConfig의 @ConditionalOnProperty(spring.data.redis.host)는 쓰지
않았다. auth가 그 프로퍼티를 갖고 있어 걸러지지 않고, test 프로파일에는 host가 없어
performance의 실앱 컨텍스트 테스트에서 설정이 사라지며, prod는 REDIS_HOST 미주입 시
스케줄러가 무증상으로 전부 멈춘다.

락 키(job-lock:{applicationName}-{profile}:{lockName})는 문자 단위로 유지한다.
조합이 달라지면 무중단 배포 중 구·신 인스턴스가 서로 다른 키를 잡아 같은 배치가
동시에 실행된다.

- spring.application.name이 비거나 없으면 기동을 실패시킨다. 중립 기본값을 두면
  이름이 빠진 서비스끼리 같은 네임스페이스를 공유해 서로의 락을 가져간다
- shedlock 의존성을 common으로 모은다. @SchedulerLock이 세 모듈 스케줄러에 직접
  붙으므로 shedlock-spring은 api로, provider는 implementation으로 둔다
- payment SchedulingConfig javadoc이 이 변경으로 거짓이 되어 함께 고친다
