+++ common/build.gradle
@@ -33,6 +33,9 @@ ext {
 // OpenFeign QueryDSL version
 def queryDslVersion = "6.10.1"
 
+// ShedLock (#439) — 락을 쓰는 서비스가 ShedLockConfig 를 @Import 한다
+def shedlockVersion = "5.13.0"
+
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
@@ -83,6 +86,12 @@ dependencies {
 
 	// redis
 	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+
+	// ShedLock — @SchedulerLock 은 락을 쓰는 서비스의 스케줄러에서 직접 참조하므로 api 로 노출한다.
+	// provider 는 ShedLockConfig 내부 구현이라 implementation 으로 둔다. 다만 이는 compileClasspath
+	// 한정이고 runtimeClasspath 로는 전 서비스에 전파된다(booking·seat·performance 가 돌려면 필요).
+	api "net.javacrumbs.shedlock:shedlock-spring:${shedlockVersion}"
+	implementation "net.javacrumbs.shedlock:shedlock-provider-redis-spring:${shedlockVersion}"
 
 	// JWT
 	implementation 'io.jsonwebtoken:jjwt-api:0.12.5'
