+++ seat-service/build.gradle
@@ -72,10 +72,6 @@ dependencies {
 
     // Redisson
     implementation 'org.redisson:redisson-spring-boot-starter:4.0.0'
-
-    // ShedLock
-    implementation 'net.javacrumbs.shedlock:shedlock-spring:5.13.0'
-    implementation 'net.javacrumbs.shedlock:shedlock-provider-redis-spring:5.13.0'
 }
 
 dependencyManagement {
